chore: update ux
