--- components/SchemaSystem/SchemaSystemHeader.vue
@@ -1,10 +1,0 @@
-<template>
-  <div
-    class="tracking-tight font-extrabold"
-  >
-    <span class="text-gray-900">Bytebase</span>
-    <span class="text-transparent bg-clip-text bg-gradient-to-br from-blue-400 to-indigo-700">
-      Schema System
-    </span>
-  </div>
-</template>
